+++ src/YoutubeAtom.stories.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { css } from '@emotion/react';
 
 import { YoutubeAtom } from './YoutubeAtom';
 import { ArticlePillar } from '@guardian/libs';
@@ -9,16 +8,15 @@ export default {
     component: YoutubeAtom,
 };
 
+const containerStyle = { width: '800px', margin: '24px' };
+const containerStyleSmall = { width: '400px', margin: '24px' };
+
 export const NoConsent = (): JSX.Element => {
     return (
-        <div
-            css={css`
-                width: 800px;
-                margin: 25px;
-            `}
-        >
+        <div style={containerStyle}>
             <YoutubeAtom
-                assetId="-ZCvZmYlQD8"
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -37,14 +35,10 @@ export const NoConsent = (): JSX.Element => {
 
 export const NoOverlay = (): JSX.Element => {
     return (
-        <div
-            css={css`
-                width: 800px;
-                margin: 25px;
-            `}
-        >
+        <div style={containerStyle}>
             <YoutubeAtom
-                assetId="-ZCvZmYlQD8"
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -65,14 +59,10 @@ export const NoOverlay = (): JSX.Element => {
 
 export const WithOverrideImage = (): JSX.Element => {
     return (
-        <div
-            css={css`
-                width: 800px;
-                margin: 25px;
-            `}
-        >
+        <div style={containerStyle}>
             <YoutubeAtom
-                assetId="3jpXAMwRSu4"
+                elementId="xyz"
+                videoId="3jpXAMwRSu4"
                 alt="Microscopic image of COVID"
                 role="inline"
                 eventEmitters={[
@@ -102,14 +92,10 @@ export const WithOverrideImage = (): JSX.Element => {
 
 export const WithPosterImage = (): JSX.Element => {
     return (
-        <div
-            css={css`
-                width: 800px;
-                margin: 25px;
-            `}
-        >
+        <div style={containerStyle}>
             <YoutubeAtom
-                assetId="N9Cgy-ke5-s"
+                elementId="xyz"
+                videoId="N9Cgy-ke5-s"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -156,14 +142,10 @@ export const WithPosterImage = (): JSX.Element => {
 
 export const WithOverlayAndPosterImage = (): JSX.Element => {
     return (
-        <div
-            css={css`
-                width: 800px;
-                margin: 25px;
-            `}
-        >
+        <div style={containerStyle}>
             <YoutubeAtom
-                assetId="N9Cgy-ke5-s"
+                elementId="xyz"
+                videoId="N9Cgy-ke5-s"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -224,14 +206,10 @@ export const GiveConsent = (): JSX.Element => {
     return (
         <>
             <button onClick={() => setConsented(true)}>Give consent</button>
-            <div
-                css={css`
-                    width: 800px;
-                    margin: 25px;
-                `}
-            >
+            <div style={containerStyle}>
                 <YoutubeAtom
-                    assetId="3jpXAMwRSu4"
+                    elementId="xyz"
+                    videoId="3jpXAMwRSu4"
                     alt="Microscopic image of COVID"
                     role="inline"
                     eventEmitters={[
@@ -267,7 +245,8 @@ export const Sticky = (): JSX.Element => {
         <div>
             <div style={{ height: '1000px' }}></div>
             <YoutubeAtom
-                assetId="-ZCvZmYlQD8"
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -292,7 +271,8 @@ export const StickyMainMedia = (): JSX.Element => {
         <div>
             <div style={{ height: '1000px' }}></div>
             <YoutubeAtom
-                assetId="-ZCvZmYlQD8"
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -312,11 +292,50 @@ export const StickyMainMedia = (): JSX.Element => {
     );
 };
 
+export const DuplicateVideos = (): JSX.Element => {
+    return (
+        <div style={containerStyleSmall}>
+            <YoutubeAtom
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
+                alt=""
+                role="inline"
+                eventEmitters={[
+                    (e) => console.log(`analytics event ${e} called`),
+                ]}
+                consentState={{}}
+                duration={252}
+                pillar={ArticlePillar.Culture}
+                height={450}
+                width={800}
+                shouldStick={true}
+            />
+            <br />
+            <YoutubeAtom
+                elementId="xyz2"
+                videoId="-ZCvZmYlQD8"
+                alt=""
+                role="inline"
+                eventEmitters={[
+                    (e) => console.log(`analytics event ${e} called`),
+                ]}
+                consentState={{}}
+                duration={252}
+                pillar={ArticlePillar.Culture}
+                height={450}
+                width={800}
+                shouldStick={true}
+            />
+        </div>
+    );
+};
+
 export const MultipleVideos = (): JSX.Element => {
     return (
         <div style={{ width: '500px' }}>
             <YoutubeAtom
-                assetId="-ZCvZmYlQD8"
+                elementId="xyz"
+                videoId="-ZCvZmYlQD8"
                 alt=""
                 role="inline"
                 eventEmitters={[
@@ -332,7 +351,8 @@ export const MultipleVideos = (): JSX.Element => {
                 title="Rayshard Brooks: US justice system treats us like 'animals'"
             />
             <YoutubeAtom
-                assetId="3jpXAMwRSu4"
+                elementId="xyu"
+                videoId="3jpXAMwRSu4"
                 alt=""
                 role="inline"
                 eventEmitters={[