+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @guardian/atoms-rendering
 
+## 23.0.0
+
+### Major Changes
+
+-   30fa8f2: YoutubeAtom duplicate video fix
+
+## 22.6.3
+
+### Patch Changes
+
+-   2f7a06a: Untsick YoutubeAtom on pause event
+-   b401b29: fix: content jump on sticky video close
+
 ## 22.6.2
 
 ### Patch Changes